Crafting issue fix, some systems are changed
Crafting issue that sometime cannot craft the backpack item and this
update includes the issue but some systems are changed.
Before that, you can open backpack after made.
Since this version, you can create ‘Not activated backpack’.
but don’t worry about it. it is easy to activate.
To activate, tap somewhere with holding not activated backpack. then,
it will be automatically activated and backpack inventory will be
displayed.